requirements.txt
Flask
Flask-Script
Flask-SQLAlchemy
WTForms